+++ tests/e2e/cases/generator_lazy_multi_yield_mutation_statement/in.ts
@@ -1,0 +1,36 @@
+let box: any = { value: 10, other: 3 };
+
+function* assignmentMutation(): Generator<string, string, any> {
+    (yield "assign-receiver")[yield "assign-key"] = 42;
+    return "assign-done";
+}
+
+function* postfixMutation(): Generator<string, string, any> {
+    (yield "postfix-receiver")[yield "postfix-key"]++;
+    return "postfix-done";
+}
+
+function* deleteMutation(): Generator<string, string, any> {
+    delete (yield "delete-receiver")[yield "delete-key"];
+    return "delete-done";
+}
+
+const assignment = assignmentMutation();
+const assignmentFirst: any = assignment.next();
+const assignmentSecond: any = assignment.next(box);
+const assignmentDone: any = assignment.next("value");
+console.log("assignment", assignmentFirst.done, assignmentFirst.value, assignmentSecond.done, assignmentSecond.value, assignmentDone.done, assignmentDone.value, box.value);
+
+box.value = 10;
+const postfix = postfixMutation();
+const postfixFirst: any = postfix.next();
+const postfixSecond: any = postfix.next(box);
+const postfixDone: any = postfix.next("value");
+console.log("postfix", postfixFirst.done, postfixFirst.value, postfixSecond.done, postfixSecond.value, postfixDone.done, postfixDone.value, box.value);
+
+box.other = 3;
+const deletion = deleteMutation();
+const deletionFirst: any = deletion.next();
+const deletionSecond: any = deletion.next(box);
+const deletionDone: any = deletion.next("other");
+console.log("delete", deletionFirst.done, deletionFirst.value, deletionSecond.done, deletionSecond.value, deletionDone.done, deletionDone.value, Object.hasOwn(box, "other"));